catch SSLError and socket.error (inc timeout) on call_galaxy
catch and log SSLErrors and socket errors and raise
a GalaxyClientAPIConnectionError exception for
main to handle and display.

Note: the default connection timeout is 20 seconds. On py2
if the connection times out, a ssl.SSLError is raised from
urls.py. On py3, a socket.timeout (socket.error subclass)
error is raised.