Parse content_spec to support indiv file/role install
When parsing the content_spec (ie, 'testing.ansible-test-content')
add support for a 'sub_name' field after the second dot in
the name that is used for selecting a specific content from the repo.

For example, for 'testing.ansible-test-content.test-role-c' the
sub_name is 'test-role-c'. For that case, 'testing' is the
namespace, ansible-test-content is the repo name, and
'test-role-c' is the specific role.

When used in combo with the '-t'/'--type' option,
this allows installing a specific role, module, or
other content.

For ex, to install the 'test-role-c' role:

 ansible-galaxy install -t role testing.ansible-test-content.test-role-c

parse_content_spec() will populate the 'sub_name' field in
it's return. I

If:
 - content_spec has two '.' in it ('a.b.c'):
    parse_content_spec() will return
    {'name': 'a.b', 'sub_name': 'c'}

    (namespace and repo name are combined into 'name')

 - content_spec has one '.' in it ('a.b'):
    parse_content_spec() will return
    {'name': 'a.b', 'sub_name': None}

 - content_spec has no '.' in it ('a'):
    parse_content_spec() will return
    {'name': 'a': 'sub_name': None}

ie, from 'testing.testing-content.elasticsearch_plugin'
'elasticsearch_plugin' is the sub_name